packages/python/bloque-ci-github/pyproject.toml
[project]
name = "bloque-ci-github"
version = "0.1.0"
description = "GitHub Actions workflow generator: CI pipelines, release/publish workflows for Python and TypeScript monorepos"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Abelardo Diaz", email = "abelardo@bloques.dev" },
]
keywords = ["github-actions", "ci", "workflow", "generator", "devops", "yaml"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Build Tools",
    "Typing :: Typed",
]
dependencies = [
    "bloque-core",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.urls]
Homepage = "https://github.com/abelardodiaz/web25-991-bloques-reciclables"
Documentation = "https://github.com/abelardodiaz/web25-991-bloques-reciclables/tree/main/packages/python/bloque-ci-github"
Repository = "https://github.com/abelardodiaz/web25-991-bloques-reciclables"
Issues = "https://github.com/abelardodiaz/web25-991-bloques-reciclables/issues"

[tool.uv.sources]
bloque-core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bloque_ci_github"]
